requirements.txt
aiogram==3.4.1
python-dotenv==1.0.0
openai==1.12.0
apscheduler==3.10.4
anthropic==0.18.1
